% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jo.R
\name{filtrer_medailles}
\alias{filtrer_medailles}
\title{Filtrer les médailles par discipline et pays}
\usage{
filtrer_medailles(data, sport, country)
}
\arguments{
\item{data}{data.frame des résultats}

\item{sport}{discipline sportive à filtrer}

\item{country}{pays (NOC) à filtrer}
}
\value{
data.frame filtré
}
\description{
Filtrer les médailles par discipline et pays
}
