% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jo.R
\name{plot_medailles_pays}
\alias{plot_medailles_pays}
\title{Graphique des médailles par pays}
\usage{
plot_medailles_pays(data, top_n = 10)
}
\arguments{
\item{data}{data.frame des résultats olympiques}

\item{top_n}{nombre de pays affichés}
}
\value{
un graphique ggplot
}
\description{
Graphique des médailles par pays
}
