% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jo.R
\name{total_medailles_pays}
\alias{total_medailles_pays}
\title{Calcul du total de médailles par pays}
\usage{
total_medailles_pays(data)
}
\arguments{
\item{data}{data.frame des résultats}
}
\value{
data.frame agrégé
}
\description{
Calcul du total de médailles par pays
}
